A fresh start for Dhallix
This commit begins Dhallix by providing types for Nix derivations.
Essentially, a Derivation is the type:

       ∀(Derivation : Type)
    → ∀(store-path : Derivation → Text)
    → ∀(Mk-Derivation : args → Derivation)
    → Derivation

Where `args` is a record type of all the properties of a derivation. The
innovation in this approach is supplying `store-path`. This function
lets us witness the output path of a derivation, but only if we're
actually in the process of forming a derivation. As any Nix/Dhallix
build is just a tower of derivations, this fits in nicely. This is the
mechanism that enables one derivation to refer to others.