Simplify the type of Derivation
The previous type of Derivation was

      ∀(Derivation : Type)
    → ∀(store-path : Derivation → Text)
    → ∀(Mk-Derivation : args → Derivation)
    → Derivation

But @Gabriel439 points out that this can be simpler. If we have (only)
store-path : Derivation → Text and Mk-Derivation : args → Derivation,
then we might as well just use the composition of these. Thus the new
type is

      ∀(Derivation : Type)
    → ∀(Mk-Derivation : args → Text)
    → Text

We may return to the old representatin if we need more operations on a
Derivation, but for now this is simpler.